requirements.txt
tabula-py
pandas